Fix: merged artifacts erroneously appeared as unloaded
Make it possible that the SquotLoadedArtifact is the working
artifact in a SquotArtifactMerge. To that end, extract the call to
#updateRightArtifacts from #buildPatch and rather call it
explicitly in cases where the right side is not given.

Curiously, something about this or the previous commit about
SquotObjectRemoval seems to have fixed the inconsistent object
graphs for file-based repositories. But not for the SquotInMemoryRepository.