WIP: Prevent earlier that the same object is walked by twice when comparing
There are some issues with the equality comparison: the = message gets
dispatched through the SqoutNameDecorator, which answers true if the two
objects have the same name. This is not what we want here.
~= is not implemented by the decorator, and thus gets forwarded to
the decorated object. The other side of the comparison is not undecorated
though, so this does not work as intended yet.

Undecorate objects before checking whether they have been reached
before. Otherwise an object could be walked by again if it appears
decorated in one place and undecorated in another place of the graph.