Patched object graphs are still inconsistent
After saving the changed Text in testTrackingText, when merging
the working copy with the actually saved snapshot to get back any
metadata changes effected by the repository, the "incoming" object
graph (i. e. that of the saved version) is inconsistent because these
two pseudo-expressions have different results:
- graph startShadow runs values at: 1
- graph shadowNamed: (graph nameOf: (startShadow runs values at: 1))
Both answer a shadow of an Array and these are equal, but not the
same. Thus, when computing the diff between the Snapshot to be
saved to the actually saved Version, the lookup of the name of that
Array fails in the incoming object graph. An addition for the Array gets
recorded where there should be none. Funnily enough, this later leads
to a "should not add merges without changes" assertion failure.

For future reference:
- The snapshot captured from the live object prior to saving is consistent.
- The base snapshot (from the previous version) is consistent.
- The right side of the patch used for saving is consistent.
- The snapshot that results from applying the patch to the snapshot
of the parent version is inconsistent. For in-memory repos this error
stays in the snapshot of the version.
	- The RunArray in `runs` is a new one, thus it gets assigned to the
	Text shadow, which also makes that shadow refer indirectly to all the
	shadows referenced by the RunArray shadow. There is also an
	ObjectAddition for the RunArray shadow, which updates the name
	mapping in the patched graph.
	- The runs values Array is also new. It therefore also has an addition
	and the name mapping gets updated.
	- But the runs values at: 1 Array is still the same, thus there is no
	object addition for it in the graph diff. Thus, nothing updates the
	mapping of its name, and the destination graph will still resolve the
	name to the old shadow. But the new RunArray shadow refers
	indirectly to an equal shadow that is not the same. Thus the graph
	will fail to look up the name of the Array in runs values at: 1 later.
	
We could avoid the problem by not applying the diffs to the shadows
again, but rather just use the right side of the already computed patch.
Unfortunately there is one piece of code where a patch is created without
a right side: SquotObjectGraphMerge privateResolvedPatch. And since
merging currently is mix-and-matching of two changesets, this would be
a place were the patching of shadow graphs would be somehow required
anyway; so it does not allow us to ignore the problem forever.