Fix comparison errors due to the = of SquotNameDecorator
Let the senders of addLeft:right:reachedVia: decide whether or
not to use this optimization, so if it turns out to be incorrect in
some cases, the = comparison can possibly be skipped.

This happens to fix testTrackedText again for practical reasons,
but the issue with inconsistent object graphs after patching still
persists.