Fix #341: Prevent that errors are hidden during commits
The ifRejected: [nil] in actionCommit turned the effective Promise
for the save into one that always gets resolved rather than rejected.

But the promise is also rejected if the user cancels the save dialog.
Now waitFor: will signal either the inner Error (or rather a copy of it)
or the BrokenPromise if it was not rejected with an Error.