Fix: incorrect package could be loaded during checkout
All packages have #() as object name, and in one more place
this caused a problem. A different package was found under
that name in the object registry.

Allow access to graphs and the object registry via mirrors. In
the future one could use two different mirrors to access both
shadows and the live objects in a SquotObjectGraph or -Registry.