Avoid expensive tree entry size computation for directory listings
To obtain the size, the object header must be read from the repository.
But most users of directory listings are not interested in the size at all
(only in the name and whether the entry is a file or a directory).

There was not even a unit test that would cover reading the size
of an entry...