Avoid flood of warnings if the repository disappeared
Warnings occur if a GitUnitOfWork is not available when the Git objects
are used. The Git Browser silently continued without a unit of work if the repository did not longer exist.

Refactor some methods a bit so that they can still do something even
without a unit of work.

Also, display only -- Repository disappeared -- and not -- Branches --
in this case. "-- Repository disappeared --" is no branch, after all.

Ref #359