Fix: modification without changes could not be applied to load order
The SquotLoadOrderMerge pretends to be an object container to get
to know the different kinds of diffs. But in the case of
SquotArtifactModification there was a shortcut that if there were no real
changes, it would just look up the existing artifact instead, which the
SquotLoadOrderMerge cannot do (because it is not a real object container
after all).

Ref #331