Fix: new Tonel packages got added with unconventional paths
Ref #366 but test case is still missing. Also there is now duplicated code
with SquotWorkingCopy>>add:at:.

In the solution of #278 in commit 54ac15f1c7b4b4c7caf6ad758ab392a2f88259fc,
the serializer for new packages gets set, but only after the package was
added and its path was validated/changed with the Cypress serializer.
Thus the packages got a Cypress path instead of a Tonel path.