Fix: error during merge when two equal package snapshots were added
For example, if you have added a package, then switch to a different branch,
which does still not have that package, the package stays in the image.
If you then add the package to the working copy again and switch back
to the branch that has the package, while keeping the uncommitted
changes, there will be an addition conflict for that package because both
sides seem to have added it independently. If you did not modify the code
in that package, the snapshots will be equal. If you then choose to keep
the already loaded edition, an error appeared: SquotDiffOfEqualObjects
does not understand: #object.

Allow SquotDiffOfEqualObjects to carry the unchanged object. It has been
a flyweight singleton so far, but keeping the polymorphy with real diffs
may be more worthwhile.

Moreover, avoid the situation by skipping diffs that do not actually have
any changes when compiling the resolved patch of a project merge.

What is curious, however, is that in the scenario described above, there is a conflict
of two added SquotPackageShadows in the SquotObjectGraphMerge, instead of a
conflict of two added SquotArtifacts in the SquotMerge. The third option would have
been that the packages are orderly merged in a SquotPackageMerge, but of the
three options the most surprising is currently the one that occurs.

Ref #365. Thanks to Christoph Thiede for discovering and reporting this.