Fix: uncommitted changes to untracked packages lost when switching branches
Situation: there are two branches, one of which tracks a certain package,
and the other one does not. The user switches from the branch that tracks
the package to the branch that does not track it. The package becomes
untracked, but it is not unloaded from the image, so the user can still 
make changes to it. Now if the user switches back to the branch that does
track the package, any modifications to that package cannot be stashed
on the branch from which the user is switching away, because it does not
track this package. Nevertheless, if "keep uncommitted changes" is used
for the switch, the modifications must not be lost.

Solution: look ahead which packages will be added to the working copy
during the switch, and save these in addition to the "official" unsaved stash.

There was an additional problem: in this situation, an addition conflict for
the package appeared (because both the branch to which the user is
switching has added the package, and the user has "added" it as unsaved
changes to the working copy), for which object graph merging was not
implemented yet. This resulted in a SquotObjectGraphMerge that had a
SquotObjectReplacement as the resolution of a 
SquotObjectAdditionConflict for the two SquotPackageShadows. But if the
SquotObjectReplacement is applied to PackageInfo, it does not have any
effect on the loaded methods. So we must rather compare the two 
versions of the package and create a SquotPackageMerge instead of the
SquotObjectReplacement. In order to do that, try to conjure an empty 
base object for the three-way merge. This works well for packages, but
may not be well-defined for other kinds of objects...

Fixes #364