Avoid loss of changes to untracked packages also during regular switch
A merge is usually unexpected during such a switch because all unsaved
changes are supposed to be saved temporarily on the branch from which
the user switches away. However, this is not sensible for packages which
are not tracked on that branch, but on the branch the user switches to.
So in order not to lose any changes to such packages, display a merge
dialog and let the user confirm which changes to restore.

Ref #364