Simplify and clarify
The load order of the working copy is already changed to be like in the
target version, and all objects that are spontaneously "added" already
exist in the target version. So each update of the temporary snapshot
in fact overwrites an object of the target version, rather than add one.
This means that the load order of the temporary snapshot is already/still
equal to the load order of the target version and no merge is required.

Ref #364