Introduce artifact ids in addition to their paths
If it is implemented in the future, this would allow to detect renames
or moves of artifacts (e. g. renaming a package, which would also
move it to a different directory).

For now, the id is stored as part of the metadata. Hide it from diff
trees in the tools so that users do not get annoyed with them.

Ids should not get changed due to merges or patches, always keep
the id that is already in the store. Except after saving a new version:
then take over the id from the version that was saved. Of course,
there already working copies without ids in existing images, and in
each image a different id will be generated at first. It is assumed
that eventually everything gets merged into the main branch and
according to the rule that the existing id is always preserved, the id
from the main branch (first parent version) will prevail over the
redundant other ids.