Fix: artificats initialized without an id could remain without an id
SquotFileStoreTest>>testPatchingPreservesId failed because the artifact written
there does not have an id in the metadata. Thus the artifact was written with an
id of nil, but it should have been written with a new UUID instead.

The rest of the tests goes on to assert that if the artifact gets written again with
a different id, that the existing id in the files is not overwritten with the new id.
This failed because nil is a special case there, to allow that older artifacts get ids
after an upgrade of Squot. Also asserting that the written id remains nil would
not make sense.