Revert "Order classes in save and merge dialogs by category"
This reverts commit b7182563aa8dec5579ae7937d5f126ce6fd19be8,
but because of the refactoring in the commit following it,
2af9f89aa57785bdafc1fb51a397858b9afb507a, this is not a clean
syntactical revert. Instead, it is just a change to get back the old behavior.

Because one cannot see the class categories in the diff tree, the ordering
of the classes by category turned out to be more surprising than helpful
at times. Since any convenient indication of the categories in the middle of
the tree view is currently not affordable, we go back to plain and simple, understandable lexicographical sorting.

Ref #363