Avoid errors when checking out versions without artifact ids
When such older versions are loaded, the artifact ids should not be removed
from the working copy. Pretend that the ids have always been there.

Ensure with the migration test that such a load preserves the id even though
it is not present in the loaded version.

Changed SquotWorkingCopy>>privateLoadVersion: to use the same method
for computing the patch as
SquotWorkingCopy>>loadVersionInteractively:ifCanceled:.