Fix: commit list incomplete when more than three commits were rebased
Several rebased commits may all have the same committer timestamp.
The heap used as a priority queue in the history walk then could not keep
the commits in the right order because they were sorted as equal.
Repeated removing and re-queuing (see VersionHistoryWalk>>atEnd
and >>next) lead to the situation where the "current" commit was
pushed back to the "end" of the heap of multiple commits with the same
timestamp. When revisiting a commit the second time, its parents may
already have been marked as visited, and were not enqueued again.
This effectively pruned the walk, leaving the shown commit list incomplete.