Do not use ClassTestCase anymore because of SUnit-pre.142
targetClass is now unwanted in favor of classToBeTested, but to
keep compatible with Squeak 5.3 we would now need to override
both selectors. Since the working copy tests are not really testing
working copy classes, but rather the combination of a repository
class with the general working copy behavior, we do not really need
the tests inherited from ClassTestCase. So stop using it and rather
use an own protocol to determine the repository class to test.