Simplify waitFor: after latest changes to Exception>>signal
A new Exception object is now created automatically if the
one that receives #signal has already been signalled before
(already has a signallerContext). This restores the desired
behavior of signalling the equivalent exception in the exception
environment (i. e. active handlers) of the waitFor: message send.
The original exception can be found in the debugger by looking
at the 'exception' temporary variable in the waitFor: context, so
that exception does not need to be passed on in the tag of the
new exception.