Fix: corrupted cached versions when excluding metadata changes
When metadata nodes in the changes chooser dialogs were added
in 2019, the implementation was quick and dirty and modified the
right side of the patch directly. When loading from older versions,
the right side is the snapshot of the version, which is cached in Squit.

Do not be destructive to either of the base version or new version
when editing the patch in the changes chooser.

Ref #302