Fix: excluding a merge did not exclude the additions&removals
Was not noticed before because of an error in SquotMergeToolTest.

Choosing the "working" version for a SquotObjectShadowMerge had
no effect except for setting the "choice" inst var, but that did not affect
the resolvedPatch in any way.

SquotObjectReferenceConflicts were created with improper
references, so the resolution SquotReferenceTargetExchange
created from it could not be applied correctly. Also, the left and right
values were initialized with the slots, rather than their values, leading
to further errors during patching.

Now the patching for conflicts resolved with the incoming edition of
a reference look up the mirror of the patched object directly. If the
owning object would rather prescribe to use a different mirror for this
subordinate object, this will not work. But to get the mirror from the
owner object, #resolvedPatch or the SquotObjectShadowMerge would
need to get to know the correct mirror, which is not implemented yet.

Refine SquotAbstractCollectionMerge>>hasChanges, so that artifact
merges with all changes excluded in the merge tool will no longer
report to have any changes. Unfortunately this requires that subclasses
override #hasChanges if they put objects in containerConflicts,
containerModifications or elementMerges that do not conform to the
conflict, diff, or merge protocols, respectively.

Remember for now: conflicts and merges respond to #hasChanges,
but diffs respond to #squotHasChanges. And the resolution of a conflict
is a diff.