Performance improvements: ref caching, object hash and size
Cache complete list of refs for prefix lookups. Previously the GitRefCache
could not know whether all references with a given prefix have already
been accessed and thus cached, or whether there are more.

Ensure that all prefix lookups include the trailing / so that the inner
methods do not have to check whether there is a / after the prefix.

Remember hash and size of loaded objects, so that the hash and size
do not need to be recalculated by writing the bytes to a dummy stream.