Show branch and tag name decorations in commit log
Further optimizations were necessary to keep the UI pause short:

Only set the label of a commit list item once, when the list is populated.
Use the string inst var inherited from PluggableListItemWrapper.

The GitObjectCache is in my repository rather full with small objects,
and GitCommits were frequently evicted because they are a bit larger
with their text, timestamps, authors. This massively hurts the comparison
with all the historians' versions. Now maintain the cache differently for
commits, so that the least recently used are evicted, not the largest.

Avoid ref name validation when creating historians for existing Git refs.

Avoid hash conversion to a string when using the cache for SquitVersions.
Explicitly clean up finalized items from that cache to get rid of the GitCommit
objects.