Speed up decorations by not searching the historians for every commit
Instead, compile a map from versions to historians once, then access
it when setting up the title of the log item.

Do not create a fake SquitVersionWrapper anymore to retrieve the
summary text. Do it the other way around: the item wrappers access
the Git Browser indirectly via their model (which is the PluggableTreeMorph),
so get the summary from the browser. The browser needs it anyway
to render the commit details.