Fix: errors when inspecting and browsing merges
The problem is that merges and conflicts handle differently than
plain diffs, but merges can also get a SquotDiffNode if there are no
conflicts. Moreover, additions and removals handle specially because
they have no original or incoming edition of the object, respectively.

The added test does not cover any of the inspect or browse actions,
but anyway there was no test for a conflict-free merge in the merge
tool yet.