Rectify license claims for code copied from Pharo
The CRC and Checksum classes were copied from Pharo in 2017.
The extension methods are also heavily based or directly copied from Pharo
during the same year. So far the copying was only documented in a
commit message from 2017, but the LICENSE file incorrectly claimed that the
code originated in the Squot project. Only the adaptations to make the code
work in Squeak originated in the Squot project. Consequently any copyright
on the original code contributed to Pharo still applies. Since Pharo no longer
keeps the authorship information from when it still used Monticello, I cannot
attribute this code more specifically than to the Pharo project as a whole.

Pharo 6 was released in June 2017 after the code was copied, that's why
the sources indicated in the LICENSE file point to Pharo 6.0.0. Also, no older
versions are available in the Pharo GitHub repository, and Pharo did not copy
the full Monticello history to their Git repository.