Add abstract copyright notice for contributors
It would be rather impractical to update the license each time when
a new person contributes something, although it would probably be the
right thing to do. The contributors' actual names can be seen in the Git
history. According to GitHub, their names until today are:
- Fabio Niephaus
- Christoph Thiede
- Patrick Rein
- Tom Beckmann
- Marcel Taeumel
- Leon Matthes
- Corinna Jaschek
- Corinna Jaschek
plus Maximilian Stiede and Leon Bein for the CoAuthors extension and
me, Jakob Reschke, as the main developer, but we three were already
mentioned by name in the license file.