Save energy and test only Trunk on Windows and macOS
Squot does not do platform-specific things or interface with plugins,
only the FileSystem API uses the FilePlugin. But FileSystem has been
split off to its own repository, so that is not a problem.

If anything in Squeak Trunk breaks on a particular platform, that is
foremost an issue to be fixed in Squeak. It is rather unlikely for
older Squeak releases to break on particular platform unless that
platform reduces backwards compatibility (such as macOS not supporting
32 bit applications anymore), in which case Squot cannot do very much.