Must also write out MCTraitDefinition if MCClassTraitDefinition is changed
Actually if it is changed, there will be one removal for the old
MCClassTraitDefinition and an addition for the new MCClassTraitDefinition.
In any case, the MCTraitDefinition must be written so that the
properties.json file in the trait directory gets updated.

Moreover there is a bug in Squeak's Monticello: when you write the
trait definition but not the corresponding class trait definition, the
"classTrait" gets trimmed away in the properties.json. The cause is
that MCTraitDefinition overrides #classTraitCompositionString incorrectly.
See MCFileTreeStCypressWriter>>#writeTraitDefinition: