Fix: there is not always an MCClassTraitDefinition in a package for every trait
For example, my FileSystem-Git package only has a MCClassTraitDefinition
for TGitActiveUnitOfWorkDuringTest, but not for TGitVisitor. A difference
is that TGitActiveUnitOfWorkDuringTest itself uses another trait, and that
also affects the class trait composition of TGitActiveUnitOfWorkDuringTest.
TGitVisitor does not use another thread, and its class trait composition is
empty. Apparently that is not worth a class trait definition...