package chooser: add menus for tracked/untracked list to control selection
While most of the time, "select none" may be of the greatest interest, the other two options "select all" and "selection inversion" fulfill two purposes: Not only offer they familiarity to the user (see also the test runner or the change list), but also might they become useful if the user is interacting in a different environment than Smalltalk where the total number of packages could be much smaller.

Note that prior to Morphic-ct.2043 (trunk), pluggable lists' menus inside a dialog can only be invoked by keyboard-focusing the list and then pressing escape.