squit: make all dialogs cancelable
This commits ensures that all senders of any #choose* or #request* selectors from SquotGUI handle the cancel case of the dialog properly. This avoids a number of confusing arbitrary nil errors.

(Note that in pure theory, something like a CancellationNotification might avoid some ugly nil checks all over the code base. However, this currently would seem overkill to me, and I am missing a sufficient understanding of promises and/or the work units architecture to think this through.)