Large refactoring: new stores concept, unified merge procedure
New implementation can be enabled with the SquotToggles useNewApi preference.
The implementation is not finished yet. Many tests are still red when the
preference is enabled. All tests should be green if it is disabled.

Because a new class for working copies is introduced in the interim, there
is a separate preference for Squit whether the GIt Browser should use
the new API (that is, display the list of "new" working copies).

A working copy does not use a single (image)store for the checked-out
objects anymore, but potentially one store per artifact. Packages use a
different store (one per Environment) than objects that are hooked up
just in the working copy, as before. In the future, stores for class variables
etc. can be introduced. This is meant to close the gap that objects need
to be hooked up somewhere in the exiting Squeak image in order to be
useful.

Unloaded objects are now simply put into a different store that does not
materialize the objects.

Likewise, storing to the file system uses different stores for the different
ways to serialize objects.

Along with this change of the stores concept comes another change:
So far, capturing, comparing, merging, patching all started with a given
object or editions thereof. This caused problems when the root object is
exchanged with another one. The new approach is that the graph
traversal starts from a reference to an object, not from that object. The
reference is fixed and determined by the store. If target/value of the
reference is exchanged, it will simply be a SquotReferenceTargetExchange,
just as if an instance variable of an object inside the graph is updated.

Artifacts no longer contain the snapshot of the object graph.
Instead, they are just items on the manifest of a project. Projects are
promoted to real objects that primarily maintain the list of artifacs. So far
the term "project" was only used in the Git Browser as a label for the list
of working copies.
To access the snapshot of object graphs, the working copy's stores must
be queried or advised to do something. SquotProjectSnapshot and
SquotArtifactSnapshot combine the mere manifest and metadata with
actual object snapshots. They also serve as caches, and replace the
previous ambiguity of "eager" SquotSnapshots and "lazy" SquotSnapshots
which capture the objects of the artifacts at different points in time.
A SquotArtifactSnapshot is always already captured. A SquotProjectSnapshot
is always lazy, except if it is based on a SquotSnapshotStore, which is
meant to hold already captured object snapshot graphs. If a project store
is backed by a working copy or file system store, it displays the behavior
of a former lazy snapshot. Working copies and file system stores can store
projects (and thus read and write artifact manifests). Other stores only
read and write artifact objects and possibly metadata, not whole projects.

SquotArtifacts shall now be value objects. Their metadata contains the id
and the path, so they are completely described by their metadata.
Artifacts as a concept may be entities, having an id, but the SquotArtifact
instances are values that represent different editions of the artifact entities.

Artifact metadata is also changed: per artifact there is now an array of
references that describe the type of the object and where to find it or
where to put it in the object memory, as well as how it should be serialized
in the file system. For the latter, it borrows from MIME media types, although
none are registered for the Smalltalk serialization formats.

All diffs should gradually be replaced by merges, because diffs are just
merges where the base version and the working copy version are the same.
Thus, SquotPatch will end up in SquotProjectMerge, which is also the
successor of SquotMerge. SquotArtifactDiffs will end up in
SquotArtifactMerge (modifications) or SquotProjectMerge (additions/removals).

Merges should no longer compute themselves (because it makes their
use in test cases more difficult), but should be computed from the outside.
Either the objects bring their own merge procedure (like
SquotTrackedObjectMetadata did before this commit) or delegate it to a
factory, builder or similar kind of service.

Merges keep a result state of the object about which they are. This is
meant to allow changes in the merge dialog to produce an edition that
does not match either of the working copy, incoming or base editions 100%.

Project merges merge artifacts that have the same id. But since artifacts
have used only the path in the past, and ids have only been introduced
recently, there is a fallback mechanism to consolidate artifacts that are
stored under the same path, but have different ids.

The merge algorithm how to traverse and object graph and react to the
differences in the references between objects are now a combination of
SquotMergeBuilder and an appropriate guide object. For object graph
merges this is SquotObjectGraphMergeGuide. For metadata merges
(previously implemented in SquotTrackedObjectMetdata directly) this is
SquotTrackedObjectMetadataMergeGuide. Project merges are implemented
in SquotProjectMergeFactory and do not yet use the unified builder, but
could follow soon.

The test strategy is to introduce abstractions or adapters for the test cases
so that the test cases can for the most part sustain both the old
implementation and the new implementation.