Treat Points and Colors as value objects
Handle them as if they are immutable.
Color is interesting because it refers to a Bitmap, which is not necessarily
a value object (although it probably is for the use in Color). So the value
object of class Color refers to a non-value object. In general this may affect
how capturing and merging must be implemented.