Fix #396: UndefinedObject errors when switching away from orphaned branch
Treat the non-existent version of the orphaned branch like an empty project.

Object merging without a base is #notYetImplemented, hence I had to resort
to a package rather than a ValueHolder in the test case.

SquotFileTestHistorian could not accept nil as a version (here for the stash
historian for unsaved changes); have it delete the historian when the version
is set to nil.