Artifact id consolidation code superseded
...by using the merge builder and the SquotArtifactReferences
that are based on the paths rather than the ids.

The method to only merge a single artifact is tricky because it cannot use
the merge builder as is because the merge builder always looks at all the
outgoing references of an object. But in order to keep the use of the callbacks
consistent, the control flow in that method is now duplicated with the merge
builder's method.