Consolidate packages with the same name during merges
Most importantly to ease the transition from older Squot versions to newer,
when ids get introduced where none existed before. Different working copies
will generate different new ids, but eventually the one on the main branch
should prevail.

In the project merge, no deserializer for packages was called yet, and the
package name was still missing in the metadata. Extend the migration from
old metadata to take the path of the artifact into account for Cypress and Tonel.