Make the order of dictionary keys in the metadata files stable
...so that the files do not change with every commit.

Turn Dictionaries into OrderedDictionaries when serializing metadata.

Ignore other variables in all kinds of dictionaries among the metadata,
so that the extra instance variables of OrderedDictionary get ignored.

Also it is not an error if singletons or numbers appear multiple times in
the same set of metadata.

Align updateMetadataAfterSaving with artifact enumeration behavior in
SquotProjectMergeFactory/SquotProjectSnapshot (effectively: use paths
instead of ids), so that it finds artifacts even if a different id was preserved
in the repository files. Now it also updates the ids in the working copy as it
was supposed to. Use putArtifact: instead of the deprecated addArtifact: there.

Change cypress metadata constructor for tests such that it conforms to the
rest of the metadata structure: the references are an array, not a dictionary.
The previous version of the constructor would make the new test green
without actually fixing the problem.