Two tests failed since omitting unchanged artifacts from merges
Unloaded artifacts are not in a special state anymore; they are just
stored in a SquotSnapshotStore instead of their otherwise appropriate
store (which would actually materialize the objects). Hence it also does
not make sense for such artifacts to appear in merges when nothing
changes about them. In a future commit, the GUI must be adapted
such that it becomes possible again to load as-yet-unloaded artifacts.