Do not register objects with the "anonymous" name
Always route the control flow through assign:to: when assigning names.
Made these changes while analyzing a strange bug that the
objectNames of the graph of the registry contained other things
than object names. It turned out that in my instance the objectNames
were the same dictionary as the objectCreators for some reason...
I did not find an explanation in the code for this. Anyway, it makes
sense to not register objects that should not have a name in the first
place, and also to check that the object names are structured as
expected.

Added bonus: some documentation about object names.