Distribute artifact migration code to the stores
Otherwise SquotTrackedObjectMetadata has to know all store implementations.
The migration for the case when a class is tracked directly (instead of a 
package) is removed for now, since this has not been implemented by a
store yet.

The SquotSmartRefStreamFile store now creates a ValueHolder reference. 
In practice, a different (yet unimplemented) kind of reference, such as a 
class variable, might actually be more appropriate. But given the information
present in an old artifact, the store cannot know or guess that. The user will
have to configure that in the working copy manually (once it becomes possible).