Fix fetching via ssh
Before, the return value of fetchFrom: was used to checkout the default branch when cloning.
Since dcc1389ccd91cab5ef6351cb1c2c571fb2f316c5, this is only needed for the internal implementation,
so the return value is not needed for the general fetchFrom: anymore. This allows us to not return the
refs when fetching externally, fixing fetching via ssh.